+++ docs/docs/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - config: add remove_request_headers @cuonglm [GH-702]
 - config: change default log level to INFO  @cuonglm [GH-902]
+- config: add pass_identity_headers @cuonglm [GH-903]
+
+### Changes
+
+- proxy: do not set X-Pomerium-Jwt-Assertion/X-Pomerium-Claim-* headers by default [GH-903]
 
 ## v0.9.1